requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
sortedcontainers
psycopg2-binary
redis
sqlalchemy
python-dotenv
pytest
requests